1. bump laravel/framework version to fix redis not found bug 2. move today caching to queue
