+++ .env.example
@@ -42,7 +42,7 @@ SESSION_DOMAIN=null
 
 BROADCAST_CONNECTION=log
 FILESYSTEM_DISK=local
-QUEUE_CONNECTION=database
+QUEUE_CONNECTION=sync
 
 CACHE_STORE=database
 CACHE_PREFIX=
