+++ app/Http/Controllers/TodayController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Jobs\CacheToday;
 use App\Repositories\Today;
-use Illuminate\Support\Facades\Artisan;
 use Inertia\Controller;
 use Inertia\Inertia;
 
@@ -11,8 +11,8 @@ class TodayController extends Controller
 {
     public function index(Today $today)
     {
-        // todo change to a job, setup docker with job workers...
-        Artisan::call('today:invalidate-cache');
+        dispatch(new CacheToday())->onQueue('default');
+
         return Inertia::render('Today', $today->toArray());
     }
