+++ composer.json
@@ -8,8 +8,9 @@
         "php": "^8.2",
         "guzzlehttp/guzzle": "^7.0",
         "inertiajs/inertia-laravel": "^1.3",
-        "laravel/framework": "^11.0",
-        "laravel/tinker": "^2.9"
+        "laravel/framework": "^11.19.0",
+        "laravel/tinker": "^2.9",
+        "predis/predis": "^2.0"
     },
     "require-dev": {
         "fakerphp/faker": "^1.23",
